configs/rock5b: remove futile unsetting of BR2_LINUX_KERNEL_INTREE_DTS_NAME
The rock5b_defconfig contained a

  # BR2_LINUX_KERNEL_INTREE_DTS_NAME is not set

line. This caused check-dotconfig.py to throw a warning in the
buildroot CI, because the explicit unsetting of the parameter is not
taken on in the actual dotconfig, but instead the dotconfig will
contain the line 'BR2_LINUX_KERNEL_INTREE_DTS_NAME=""'.

This patch removes the parameter from the rock5b_defconfig. The
resulting dotconfig from the original rock5b_defconfig and the
rock5b_defconfig without the parameter is identical, but the
check-dotconfig.py does not throw a warning anymore.

Signed-off-by: Kilian Zinnecker <kilian.zinnecker@mail.de>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>