package/multipath-tools: drop README.md from license files
Drop README.md from license files as advocated by Arnout Vandecappelle:
https://patchwork.ozlabs.org/project/buildroot/patch/20240314211117.9103-1-fontaine.fabrice@gmail.com

The only license-related part of README.md is that specifies that the
default license, if not mentioned in the file, is LGPL-2.0. Since the
README file is likely to be updated with every new release, this is a
lot of overhead for such a small added value.

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>