configs/nezha_defconfig: use mainline Linux 6.6.22
The D1 support is now in mainline, so use that instead.  There is no
dedicated nezha defconfig, so use the riscv defconfig.  This defconfig has
most drivers as modules, so add mdev to ensure they get correctly loaded.

The defconfig does not have USB gadget/OTG support, so enable that using a
config fragment to make the USB-C connector work and get rid of the
following warning:

[    7.233418] musb-sunxi 4100000.usb: Invalid or missing 'dr_mode' property
[    7.240330] musb-sunxi: probe of 4100000.usb failed with error -22

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>