package/googlefontdirectory: fix command to generate license file hashes
In commit ebe5d9edfe8d07eafd7b61c4f77ce509bb98a0ad ("boot, package,
support, toolchain: switch to 2 spaces for the hash file"), all hash
files were migrated to use 2 spaces as the separator.

However, in the googlefontdirectory hash file, a command is present in
a comment to indicate how to generate the part of the hash file that
provides the list of license file hashes. This command was not updated
as part of ebe5d9edfe8d07eafd7b61c4f77ce509bb98a0ad, so it still emits
a result in which a single space is used a separator between the hash
type (sha256) and the hash value.

This commit fixes that by using a 2-space separator.

Signed-off-by: Christian Hitz <christian.hitz@bbv.ch>
[Thomas: extracted from
https://patchwork.ozlabs.org/project/buildroot/patch/20240228145013.411919-2-christian@klarinett.li/
into a separate patch]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>