package/googlefontdirectory: sort license file hashes
The current command listed in the hash file to generate the list of
license files hashes does not sort files. Due to this, depending on
who runs the command, the result can potentially be different.

In order to address this, this commit changes the command to sort
license files, and updates the hash file accordingly.

Signed-off-by: Christian Hitz <christian.hitz@bbv.ch>
[Thomas: extract from
https://patchwork.ozlabs.org/project/buildroot/patch/20240228145013.411919-2-christian@klarinett.li/
into a separate patch]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
(cherry picked from commit 7b355f5d9801e8d215fba864a33e6bf62623eda0)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>