package/googlefontdirectory: bump to new version
This is the current version on main branch.

For Apache-2.0 licensed fonts, there are new fonts, and fonts that
have been removed.

For OFL-1.0 fonts, there are new fonts, fonts removed, but also fonts
where the license file has changed. In this latter case, the changes
are http:// to https:// in some URLs, copyright year changes, a minor
update of the OFL license (from Feb 1, 2007 to Feb 26, 2007, which has
a small rewording), text rewrapping.

Signed-off-by: Christian Hitz <christian.hitz@bbv.ch>
[Thomas: extracted from
https://patchwork.ozlabs.org/project/buildroot/patch/20240228145013.411919-2-christian@klarinett.li/,
added details about license file changes]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>