package/luaposix: fix build with glibc 2.39
When glibc was bumped to version 2.39 in commit
b5680f53d60acf8ff6010082f873438a39bd5d97 it removed the deprecated
libcrypt support.

As glibc's libcrypt was providing luaposix's libcrypt dependency this
broke any luaposix build using glibc version 2.39.

To fix this select the libxcrypt dependency to luaposix when using a
glibc toolchain and add the dependency if selected.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
[yann.morin.1998@free.fr: use more traditional ifeq-endif block]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>