package/dovecot: add optional dependency to libxcrypt
When glibc was bumped to version 2.39 in commit
b5680f53d60acf8ff6010082f873438a39bd5d97 it removed the deprecated
libcrypt support.

As glibc's libcrypt was providing dovecot's libcrypt dependency this
broke the dovecot build using glibc version 2.39.

To fix this select the libxcrypt dependency to dovecot when using a
glibc toolchain and add the dependency if selected.

Fixes:
http://autobuild.buildroot.net/results/99d816fb7a033ec73a958200041f0dcd728e0e87/

Signed-off-by: Bernd Kuhls <bernd@kuhls.net>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>