package/leafnode2: add optional dependency to libxcrypt
When glibc was bumped to version 2.39 in commit
b5680f53d60acf8ff6010082f873438a39bd5d97 it removed the deprecated
libcrypt support.

As glibc's libcrypt was providing leafnode2's libcrypt dependency this
broke the leafnode2 build using glibc version 2.39.

To fix this select the libxcrypt dependency to leafnode2 when using a
glibc toolchain and add the dependency if selected.

Fixes:
http://autobuild.buildroot.net/results/08f86ce128623105087cda0c5b32a1d0d0cfa166/

Signed-off-by: Bernd Kuhls <bernd@kuhls.net>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>