package/lighttpd: add optional dependency to libxcrypt
When glibc was bumped to version 2.39 in commit
b5680f53d60acf8ff6010082f873438a39bd5d97 it removed the deprecated
libcrypt support.

As glibc's libcrypt was providing lighttpd's libcrypt dependency this
broke the lighttpd build using glibc version 2.39.

To fix this select the libxcrypt dependency to lighttpd when using a
glibc toolchain and add the dependency if selected.

Signed-off-by: Bernd Kuhls <bernd@kuhls.net>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>