package/tiff: explicitly disable jbig support for host and target
BR2_PACKAGE_TIFF_JBIG did actually not do anything, as no explicit
--enable-jbig was passed to configure and there is no libjbig in Buildroot,
so drop it and instead explicitly disable jbig support.

Also add --disable-jbig for the host build, which was missed when host
support was added in commit 91b16fbbf9db (tiff: add host variant).

As the TIFF_JBIG option was a noop, do not add legacy handling for it.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>