configs/warpboard: remove defconfig
As suggested by Fabio in [1], this commit removes the defconfig and all
other associated files.

The U-Boot support for the Warpboard has been removed upstream in
commit [2], first included in version 2022.10.

For reference, at the time of this commit, the latest Kernel is 6.8.6
and still has support for this board. See [3].

The reason of this removal is because this defconfig fail to build with
recent toolchains. See [4].

Fixes: [4]

[1] https://lists.buildroot.org/pipermail/buildroot/2024-April/689596.html
[2] https://source.denx.de/u-boot/u-boot/-/commit/c50ff4a933d5617415221bbb85b0bfa5247b697a
[3] https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/tree/arch/arm/boot/dts/nxp/imx/imx6sl-warp.dts?h=v6.8.6
[4] https://gitlab.com/buildroot.org/buildroot/-/jobs/6609466749

Suggested-by: Fabio Estevam <festevam@gmail.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>
Reviewed-by: Fabio Estevam <festevam@gmail.com>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>